backend/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography==41.0.7
alembic==1.13.0

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pyotp==2.9.0
qrcode==7.4.2
pillow==10.1.0

# Proxmox
proxmoxer==2.0.1
requests==2.31.0

# SSH and remote management
paramiko==3.4.0

# Configuration and environment
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# YAML processing for cloud-init
pyyaml==6.0.1

# Logging
python-json-logger==2.0.7

# File handling
aiofiles==23.2.1

# Validation
email-validator==2.1.0

# Testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
weasyprint==66.0
markdown==3.10
